refactor: cardType 관련 로직 Utils로 분리
